+++ agent/src/agent.php
@@ -4,6 +4,7 @@ namespace Laravel\NightwatchAgent;
 
 use Closure;
 use Laravel\NightwatchAgent\Contracts\Browser;
+use Laravel\NightwatchAgent\Contracts\Clock as ClockContract;
 use Laravel\NightwatchAgent\Factories\BrowserFactory;
 use Psr\Http\Message\ResponseInterface;
 use React\EventLoop\Loop as BaseLoop;
@@ -11,6 +12,7 @@ use React\EventLoop\LoopInterface;
 use React\EventLoop\StreamSelectLoop;
 use React\Socket\ServerInterface;
 use React\Socket\TcpServer;
+use React\Stream\WritableResourceStream;
 
 use function date;
 use function file_get_contents;
@@ -37,6 +39,8 @@ $browserFactory ??= null;
 $serverResolver ??= null;
 /** @var ?LoopInterface $loop */
 $loop ??= null;
+/** @var ?ClockContract $clock */
+$clock ??= null;
 
 /*
  * Input...
@@ -79,8 +83,13 @@ BaseLoop::set($loop);
  * Logging helpers...
  */
 
-$stdOut = new OutputWriter($loop, STDOUT);
-$stdErr = new OutputWriter($loop, STDERR);
+[$asyncStdOut, $asyncStdError] = match (PHP_OS_FAMILY) {
+    'Windows' => [null, null],
+    default => [new WritableResourceStream(STDOUT), new WritableResourceStream(STDERR)],
+};
+
+$stdOut = new OutputWriter($loop, syncStream: STDOUT, asyncStream: $asyncStdOut);
+$stdErr = new OutputWriter($loop, syncStream: STDERR, asyncStream: $asyncStdError);
 
 $debug = static function (string $message) use ($verbose, $stdOut): void {
     if ($verbose) {
@@ -130,6 +139,8 @@ $packageVersion = rtrim(file_get_contents($basePath.'/../../version.txt') ?: '')
 /*
  * Initialize services...
  */
+$clock ??= new Clock;
+
 $browserFactory ??= new BrowserFactory;
 
 $ingestDetailsBrowser = $browserFactory(
@@ -148,6 +159,7 @@ $ingestDetailsBrowser = $browserFactory(
 $ingestDetails = new IngestDetailsRepository(
     loop: $loop,
     browser: $ingestDetailsBrowser,
+    clock: $clock,
     onAuthenticationSuccess: static fn (IngestDetails $ingestDetails, float $duration) => $info('Authentication successful ['.round($duration, 3).'s]'),
     onAuthenticationError: static fn (string $message, float $duration) => $error('Authentication failed ['.round($duration, 3).'s]: '.$message),
     onUnderQuota: static function () use (&$ingest) {
@@ -171,6 +183,7 @@ $ingest = new Ingest(
     loop: $loop,
     browser: $ingestBrowser,
     ingestDetails: $ingestDetails,
+    clock: $clock,
     buffer: new StreamBuffer(6_000_000),
     concurrentRequestLimit: 2,
     maxBufferDurationInSeconds: 10,